+++ src/main/java/mirror/normalasm/core/NormalTransformer.java
@@ -142,8 +142,7 @@ public class NormalTransformer implements IClassTransformer {
         if (NormalConfig.instance.fixMC31681) {
             addTransformation("net.minecraft.client.renderer.EntityRenderer", this::fixMC31681);
         }
-        addTransformation("net.minecraft.nbt.NBTTagCompound", bytes -> nbtTagCompound$replaceDefaultHashMap(bytes, NormalConfig.instance.optimizeNBTTagCompoundBackingMap, NormalConfig.instance.nbtBackingMapStringCanonicalization, NormalConfig.instance.optimizeNBTTagCompoundBackingMapOpenMap));    }
-
+        addTransformation("net.minecraft.nbt.NBTTagCompound", bytes -> nbtTagCompound$replaceDefaultHashMap(bytes, NormalConfig.instance.optimizeNBTTagCompoundBackingMap, NormalConfig.instance.optimizeNBTTagCompoundMapThreshold, NormalConfig.instance.nbtBackingMapStringCanonicalization));}
     public void addTransformation(String key, Function<byte[], byte[]> value) {
         NormalLogger.instance.info("Adding class {} to the transformation queue", key);
         transformations.put(key, value);
@@ -466,8 +465,8 @@ public class NormalTransformer implements IClassTransformer {
         return writer.toByteArray();
     }
 
-    private byte[] nbtTagCompound$replaceDefaultHashMap(byte[] bytes, boolean optimizeMap, boolean canonicalizeString, boolean useOpenMap) {
-        if (!optimizeMap && !canonicalizeString) {
+    private byte[] nbtTagCompound$replaceDefaultHashMap(byte[] bytes, boolean optimizeMap, int mapThreshold, boolean canonicalizeString) {
+        if ((!optimizeMap || mapThreshold == 0) && !canonicalizeString) {
             return bytes;
         }
         ClassReader reader = new ClassReader(bytes);
@@ -480,12 +479,12 @@ public class NormalTransformer implements IClassTransformer {
                 while (iter.hasNext()) {
                     AbstractInsnNode instruction = iter.next();
                     if (instruction.getOpcode() == INVOKESTATIC) {
-                        iter.set(new TypeInsnNode(NEW, useOpenMap ? (canonicalizeString ? "mirror/normalasm/api/datastructures/canonical/AutoCanonizingOpenHashMap" : "it/unimi/dsi/fastutil/objects/Object2ObjectOpenHashMap") : (canonicalizeString ? "mirror/normalasm/api/datastructures/canonical/AutoCanonizingArrayMap" : "it/unimi/dsi/fastutil/objects/Object2ObjectArrayMap")));
-                        iter.add(new InsnNode(DUP));
-                        iter.add(new MethodInsnNode(INVOKESPECIAL, useOpenMap ? (canonicalizeString ? "mirror/normalasm/api/datastructures/canonical/AutoCanonizingOpenHashMap" : "it/unimi/dsi/fastutil/objects/Object2ObjectOpenHashMap") : (canonicalizeString ? "mirror/normalasm/api/datastructures/canonical/AutoCanonizingArrayMap" : "it/unimi/dsi/fastutil/objects/Object2ObjectArrayMap"), "<init>", "()V", false));
+                        iter.set(new TypeInsnNode(NEW, "mirror/normalasm/api/datastructures/TagMap"));                        iter.add(new InsnNode(DUP));
+                        iter.add(new MethodInsnNode(INVOKESPECIAL, "mirror/normalasm/api/datastructures/TagMap", "<init>", "()V", false));
                         break;
                     }
                 }
+                break;
             }
         }